src/requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0

# Database
asyncpg>=0.29.0

# Image Generation (Phase 4)
Pillow>=10.0.0

# Optional but recommended
aiohttp>=3.9.0

numpy>=2.3.3
